// src/openapi/client/models/PlaceBase.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { PlaceLevelChoices } from './PlaceLevelChoices';
export type PlaceBase = {
    slug?: (string | null);
    locationLevel?: (PlaceLevelChoices | null);
    isActive?: (boolean | null);
    parentId?: (number | null);
};